Handing of groups when running tests
The service configuration files incorrectly obtained the groups to use
when starting up. If multiple groups were desired, the parsing approach
would fail. Additonally in the test management configuration, the log
directory was obtained by getting the groups system property. In genral
this workds, but when multple groups were used the corresponding log
directory was be the combined group names.